data-transform-to-bq/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cramjam==2.3.2
fastparquet==0.6.3
fsspec==2021.6.0
gcsfs
google-api-core==1.30.0
google-auth==1.30.2
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1
google-cloud-storage==1.13.1
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
idna==2.10
numpy==1.20.3
packaging==20.9
pandas==1.2.4
protobuf==3.17.3
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
thrift==0.13.0
urllib3==1.26.5
SQLAlchemy==1.4.1     
pg8000==1.19.5
google-cloud-secret-manager==1.0.0